+++ ntex/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [2.8.0] - 2024-11-04
+
+* Use updated Service trait
+
+* Don't swallow error when calling panic for read_response_json #443
+
 ## [2.7.0] - 2024-10-16
 
 * Better handling for h2 remote payload